+++ src/components/molecule/tabs/Tabs.js
@@ -2,22 +2,33 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Tabs } from 'react-tabs';
 
+import classNames from 'classnames';
+
 import css from './Tabs.module.scss';
 
-const TabsComponent = ({ children, ...otherProps }) => {
+const TabsComponent = ({ dark, children, ...otherProps }) => {
+  const classes = classNames({
+    [css['c-tabs']]: true,
+    [css['c-tabs--dark']]: !!dark
+  });
+
   return (
-    <Tabs className={css['c-tabs']} {...otherProps}>
+    <Tabs className={classes} {...otherProps}>
       {children}
     </Tabs>
   );
 };
 
 TabsComponent.propTypes = {
-  children: PropTypes.oneOfType([PropTypes.object, PropTypes.array])
+  children: PropTypes.oneOfType([PropTypes.object, PropTypes.array]),
+  dark: PropTypes.bool
 };
 
 TabsComponent.defaultProps = {
-  children: null
+  children: null,
+  dark: false
 };
+
+TabsComponent.displayName = 'TabsComponent';
 
 export default TabsComponent;